requirements.txt
discord.py-self==1.9.2
colorama
pypresence
pillow
numpy